Add Selection Strategy in Sample
